Refactored things a bit to remove a forEach loop
