+++ plugins/Profiler/views/SnapshotRanked.js
@@ -19,7 +19,7 @@ import AutoSizer from 'react-virtualized-auto-sizer';
 import { FixedSizeList as List } from 'react-window';
 import ChartNode from './ChartNode';
 import NoSnapshotDataMessage from './NoSnapshotDataMessage';
-import { barHeight, computeSelfDurations, getGradientColor, minBarWidth, scale } from './constants';
+import { barHeight, calculateSelfDuration, getGradientColor, minBarWidth, scale } from './constants';
 
 type Node = {|
   id: any,
@@ -245,7 +245,7 @@ const getNodeIndex = memoize((rankedData: RankedData, id: string | null): number
 });
 
 const convertSnapshotToChartData = (snapshot: Snapshot, showNativeNodes: boolean): RankedData => {
-  const [selfDurations, maxSelfDuration] = computeSelfDurations(snapshot, showNativeNodes);
+  let maxSelfDuration = 0;
 
   const nodes = snapshot.committedNodes
     .filter(nodeID => {
@@ -254,9 +254,10 @@ const convertSnapshotToChartData = (snapshot: Snapshot, showNativeNodes: boolean
       return (nodeType === 'Composite' || (nodeType === 'Native' && showNativeNodes));
     })
     .map((nodeID, index) => {
-      const node = snapshot.nodes.get(nodeID);
-      const name = node.get('name') || 'Unknown';
-      const selfDuration = selfDurations[nodeID] || 0;
+      const selfDuration = calculateSelfDuration(snapshot, nodeID);
+      maxSelfDuration = Math.max(maxSelfDuration, selfDuration);
+
+      const name = snapshot.nodes.getIn([nodeID, 'name']) || 'Unknown';
       const label = `${name} (${selfDuration.toFixed(1)}ms)`;
       return {
         id: nodeID,
